s3-scan/requirements.txt
pyClamd==0.4.0
python-dotenv==0.10.3
